Use newer appimage tool for the appimage creation part Strip binaries using system strip but not for snapshot AppImage Add --needed to arch install command Add ffi and icu libs for all distros Check submodules are checked out and if not check them out Add developer lines to README

+++ installers/appimage/create_appimage.sh
@@ -9,6 +9,11 @@ run_script() {
     source ../../DevEnvironments/Linux/os_detection.sh
     source ../../DevEnvironments/Linux/install_package_fn.sh
     source ../../DevEnvironments/common/check_rename_fn.sh
+
+    if ! install_package fuse; then
+        return 1
+    fi
+
     APPDIR="Oolite.AppDir"
     rm -rf $APPDIR
     mkdir -p $APPDIR/usr/bin
@@ -20,23 +25,38 @@ run_script() {
     curl -L -O https://github.com/linuxdeploy/linuxdeploy/releases/download/continuous/linuxdeploy-x86_64.AppImage
     chmod +x linuxdeploy-x86_64.AppImage
 
-    if ! install_package fuse; then
-        return 1
-    fi
-
     case "$CURRENT_DISTRO" in
         debian) SDL2="--library=/usr/lib/x86_64-linux-gnu/libSDL2-2.0.so.0" ;;
         redhat) SDL2="--library=/usr/lib64/libSDL2-2.0.so.0 --library=/usr/lib64/libSDL3.so.0" ;;
         arch) SDL2="--library=/usr/lib/libSDL2-2.0.so.0 --library=/usr/lib/libSDL3.so.0" ;;
     esac
 
+    echo "Building AppDir for AppImage..."
     if ! NO_STRIP=1 ./linuxdeploy-x86_64.AppImage \
     --appdir $APPDIR \
     --executable $PROGDIR/oolite \
     --desktop-file ../FreeDesktop/oolite.desktop \
     --icon-file ../FreeDesktop/oolite-icon.png \
-    $SDL2 \
-    --output appimage; then
+    $SDL2; then
+        echo "❌ AppDir generation failed!" >&2
+        return 1
+    fi
+
+   	if (( $# == 2 )) && [[ $2 == "dev" ]]; then
+        echo "Not stripping libs for snapshot AppImage"
+   	else
+        echo "Stripping libs in AppDir..."
+        find "$APPDIR/usr" -type f \
+        \( -name '*.so' -o -name '*.so.*' \) \
+        -exec strip --strip-unneeded '{}' +   # keeps symbols needed for runtime linking
+    fi
+
+    rm -f appimagetool-x86_64.AppImage
+    curl -L -O https://github.com/AppImage/appimagetool/releases/download/continuous/appimagetool-x86_64.AppImage
+    chmod +x appimagetool-x86_64.AppImage
+
+    echo "Creating AppImage..."
+    if ! ./appimagetool-x86_64.AppImage $APPDIR; then
         echo "❌ AppImage creation failed!" >&2
         return 1
     fi
